refactor(platform): improve readability and maintainability
