+++ apps/platform/src/components/auth/index.ts
@@ -1,0 +1,7 @@
+import AuthHeader from '@/components/auth/auth-header'
+import AuthErrorBanner from '@/components/auth/auth-error-banner'
+import AuthSocialGrid from '@/components/auth/auth-social-grid'
+import AuthEmailForm from '@/components/auth/auth-email-form'
+import AuthTnC from '@/components/auth/auth-tnc'
+
+export { AuthHeader, AuthErrorBanner, AuthSocialGrid, AuthEmailForm, AuthTnC }
