+++ apps/platform/src/components/auth/otp/index.ts
@@ -1,0 +1,5 @@
+import OtpHeader from '@/components/auth/otp/otp-header'
+import OtpInputForm from '@/components/auth/otp/otp-input-form'
+import OtpResend from '@/components/auth/otp/otp-resend'
+
+export { OtpHeader, OtpInputForm, OtpResend }
