+++ apps/platform/src/components/shared/sidebar/index.tsx
@@ -14,6 +14,7 @@ import TierLimit from './tierLimit'
 import { Combobox } from '@/components/ui/combobox'
 import { selectedWorkspaceAtom } from '@/store'
 import { Button } from '@/components/ui/button'
+import { TIRE_PLAN, VERSION_BADGE } from '@/constants/sidebar'
 
 function Sidebar(): JSX.Element {
   const selectedWorkspace = useAtomValue(selectedWorkspaceAtom)
@@ -59,7 +60,7 @@ function Sidebar(): JSX.Element {
             <KeyshadeLogoSVG /> Keyshade
           </div>
           <div className="rounded bg-white/10 px-2 py-[0.12rem] text-xs font-bold">
-            ALPHA
+            {VERSION_BADGE}
           </div>
         </div>
         <Combobox />
@@ -79,7 +80,7 @@ function Sidebar(): JSX.Element {
         <div className="absolute bottom-10 w-[17rem] rounded-lg border border-white/10 bg-white/5 p-4">
           <div className="mb-5 flex items-center gap-3">
             <Button className="h-6 bg-[#60A5FA4D] p-3 text-white hover:bg-[#60A5FA4D]">
-              Free
+              {TIRE_PLAN}
             </Button>
           </div>
