+++ apps/platform/src/constants/sidebar/index.ts
@@ -1,0 +1,2 @@
+export const TIRE_PLAN = "Free"
+export const VERSION_BADGE = "ALPHA";
